task3_1P/src/main/java/sit707_tasks/DateUtil.java
package sit707_tasks;

public class DateUtil {
    private int day;
    private int month;
    private int year;

    public DateUtil(int day, int month, int year) {
        if (!isValidDate(day, month, year)) {
            throw new IllegalArgumentException("Invalid date: " + day + "/" + month + "/" + year);
        }
        this.day = day;
        this.month = month;
        this.year = year;
    }

    private boolean isValidDate(int day, int month, int year) {
        int maxDay = getMaxDayOfMonth(month, year);
        return day >= 1 && day <= maxDay && month >= 1 && month <= 12;
    }

    private int getMaxDayOfMonth(int month, int year) {
        switch (month) {
            case 2:
                return isLeapYear(year) ? 29 : 28;
            case 4: case 6: case 9: case 11:
                return 30;
            default:
                return 31;
        }
    }

    private boolean isLeapYear(int year) {
        return (year % 4 == 0 && year % 100 != 0) || (year % 400 == 0);
    }

    public void increment() {
        if (day < getMaxDayOfMonth(month, year)) {
            day++;
        } else {
            day = 1;
            if (month == 12) {
                month = 1;
                year++;
            } else {
                month++;
            }
        }
    }

    public void decrement() {
        if (day > 1) {
            day--;
        } else {
            if (month == 1) {
                month = 12;
                day = 31;
                year--;
            } else {
                month--;
                day = getMaxDayOfMonth(month, year);
            }
        }
    }

    public int getDay() {
        return day;
    }

    public int getMonth() {
        return month;
    }

    public int getYear() {
        return year;
    }
}
